requirements.txt
flask                   ==  2.0.3
flask-mysql             ==  1.5.2
flask-sqlalchemy        ==  2.5.1
flask-restful           ==  0.3.9
flask-marshmallow       ==  0.14.0
flask-cors              ==  3.0.10
pymysql                 ==  1.0.2
mysqlclient             ==  2.1.0
marshmallow-sqlalchemy  ==  0.28.0
mariadb                 ==  1.0.10
cryptography            ==  36.0.1
pyjwt                   ==  2.3.0
itsdangerous            ==  2.1.1
gunicorn                ==  20.1.0
setproctitle            ==  1.2.2
sendgrid                ==  6.9.7